remove network state handling from EnterAmountViewModel
